02/09/2022

+++ SkyExams/Controllers/Lesson_PlanController.cs
@@ -23,129 +23,55 @@ namespace SkyExams.Controllers
 
         public ActionResult planScreen(int? id)
         {
-            Sys_User user = db.Sys_User.ToList().Find(u => u.SysUser_ID == id);
-            return View(user);
+            ViewData["userID"] = "" + id;
+            Sys_User forRole = db.Sys_User.ToList().Find(u => u.SysUser_ID == id);
+            ViewData["userRole"] = "" + forRole.User_Role_ID;
+            List<Plane_Type> planeTypes = db.Plane_Type.ToList();
+
+            return View(planeTypes);
         }// returns lesson plan screen
+
+        public FileContentResult getImg(int id)
+        {
+            byte[] byteArray = db.Plane_Type.Find(id).Plane_Image;
+            return byteArray != null
+                ? new FileContentResult(byteArray, "image/jpeg")
+                : null;
+        }
 
         public ActionResult topicScreen(int? id, int? topicId)
         {
             ViewData["userID"] = "" + id;
             Sys_User user = db.Sys_User.Find(id);
+            ViewData["userRole"] = "" + user.User_Role_ID;
             int userRole = Convert.ToInt32(user.User_Role_ID);
             List<Lesson_Plan> planList = new List<Lesson_Plan>();
             if(userRole == 1)
             {
-                if (topicId == 1)
+                ViewData["topicId"] = "" + topicId;
+                ViewData["planeType"] = db.Plane_Type.ToList().Find(p => p.Plane_Type_ID == topicId).Type_Description;
+                Student tempStu = db.Students.ToList().Find(s => s.SysUser_ID == user.SysUser_ID);
+                List<Student_Lesson_Plan> stuPlan = db.Student_Lesson_Plan.ToList().FindAll(r => r.Student_ID == tempStu.Student_ID);
+                foreach (Student_Lesson_Plan temp in stuPlan)
                 {
-                    ViewData["topicId"] = "" + 1;
-                    ViewData["planeType"] = "Cessna 172";
-                    List<Student_Lesson_Plan> stuPlan = db.Student_Lesson_Plan.ToList().FindAll(r => r.Student_ID == user.SysUser_ID);
-                    foreach (Student_Lesson_Plan temp in stuPlan)
+                    List<Lesson_Plan> tempPlans = db.Lesson_Plan.ToList();
+                    foreach (Lesson_Plan temPlan in tempPlans)
                     {
-                        List<Lesson_Plan> tempPlans = db.Lesson_Plan.ToList();
-                        foreach (Lesson_Plan temPlan in tempPlans)
+                        if (temPlan.Lesson_Plan_ID == temp.Lesson_Plan_ID && temPlan.Rating_ID == topicId)
                         {
-                            if (temPlan.Lesson_Plan_ID == temp.Lesson_Plan_ID && temPlan.Topic_ID == 1)
-                            {
-                                planList.Add(temPlan);
-                            }
-                        }// inner for each
-                    }// for each
-                    return View(planList);
-                }// cessna 172
-                if (topicId == 2)
-                {
-                    ViewData["topicId"] = "" + 2;
-                    ViewData["planeType"] = "Cessna 172 RG";
-                    List<Student_Lesson_Plan> stuPlan = db.Student_Lesson_Plan.ToList().FindAll(r => r.Student_ID == user.SysUser_ID);
-                    foreach (Student_Lesson_Plan temp in stuPlan)
-                    {
-                        List<Lesson_Plan> tempPlans = db.Lesson_Plan.ToList();
-                        foreach (Lesson_Plan temPlan in tempPlans)
-                        {
-                            if (temPlan.Lesson_Plan_ID == temp.Lesson_Plan_ID && temPlan.Topic_ID == 2)
-                            {
-                                planList.Add(temPlan);
-                            }
-                        }// inner for each
-                    }// for each
-                    return View(planList);
-                }// cessna 172 RG
-                if (topicId == 3)
-                {
-                    ViewData["topicId"] = "" + 3;
-                    ViewData["planeType"] = "Cherokee 140";
-                    List<Student_Lesson_Plan> stuPlan = db.Student_Lesson_Plan.ToList().FindAll(r => r.Student_ID == user.SysUser_ID);
-                    foreach (Student_Lesson_Plan temp in stuPlan)
-                    {
-                        List<Lesson_Plan> tempPlans = db.Lesson_Plan.ToList();
-                        foreach (Lesson_Plan temPlan in tempPlans)
-                        {
-                            if (temPlan.Lesson_Plan_ID == temp.Lesson_Plan_ID && temPlan.Topic_ID == 3)
-                            {
-                                planList.Add(temPlan);
-                            }
-                        }// inner for each
-                    }// for each
-                    return View(planList);
-                }// cherokee 140
-                if (topicId == 4)
-                {
-                    ViewData["topicId"] = "" + 4;
-                    ViewData["planeType"] = "Piper Twin Comanche";
-                    List<Student_Lesson_Plan> stuPlan = db.Student_Lesson_Plan.ToList().FindAll(r => r.Student_ID == user.SysUser_ID);
-                    foreach (Student_Lesson_Plan temp in stuPlan)
-                    {
-                        List<Lesson_Plan> tempPlans = db.Lesson_Plan.ToList();
-                        foreach (Lesson_Plan temPlan in tempPlans)
-                        {
-                            if (temPlan.Lesson_Plan_ID == temp.Lesson_Plan_ID && temPlan.Topic_ID == 4)
-                            {
-                                planList.Add(temPlan);
-                            }
-                        }// inner for each
-                    }// for each
-                    return View(planList);
-                }// cessna 172
-                else
-                {
-                    return View(planList);
-                }
+                            planList.Add(temPlan);
+                        }
+                    }// inner for each
+                }// for each
+                return View(planList);
+                
             }// if user is a student
             else
             {
-                if (topicId == 1)
-                {
-                    ViewData["topicID"] = "" + 1;
-                    ViewData["planeType"] = "Cessna 172";
-                    planList = db.Lesson_Plan.ToList().FindAll(p => p.Topic_ID == 1);
-                    return View(planList);
-                }// cessna 172
-                if (topicId == 2)
-                {
-                    ViewData["topicID"] = "" + 2;
-                    ViewData["planeType"] = "Cessna 172 RG";
-                    planList = db.Lesson_Plan.ToList().FindAll(p => p.Topic_ID == 2);
-                    return View(planList);
-                }// cessna 172 RG
-                if (topicId == 3)
-                {
-                    ViewData["topicID"] = "" + 3;
-                    ViewData["planeType"] = "Cherokee 140";
-                    planList = db.Lesson_Plan.ToList().FindAll(p => p.Topic_ID == 3);
-                    return View(planList);
-                }// cherokee 140
-                if (topicId == 4)
-                {
-                    ViewData["topicID"] = "" + 4;
-                    ViewData["planeType"] = "Piper Twin Comanche";
-                    planList = db.Lesson_Plan.ToList().FindAll(p => p.Topic_ID == 4);
-                    return View(planList);
-                }// twin commanche
-                else
-                {
-                    return View(planList);
-                }
+                ViewData["topicID"] = "" + topicId;
+                ViewData["planeType"] = db.Plane_Type.ToList().Find(p => p.Plane_Type_ID == topicId).Type_Description;
+                planList = db.Lesson_Plan.ToList().FindAll(p => p.Rating_ID == topicId);
+                return View(planList);
             }// else
         }// theme screen
 
@@ -179,8 +105,8 @@ namespace SkyExams.Controllers
                 newPlan.Lesson_Plan_ID = resourceId;
                 newPlan.LP_Name = name;
                 int theme = Convert.ToInt32(topicId);
-                newPlan.Topic_ID = theme;
-                int instructorId = Convert.ToInt32(id);
+                newPlan.Rating_ID = theme;
+                int instructorId = db.Instructors.ToList().Find(i => i.SysUser_ID == id).Instructor_ID;
                 newPlan.Instructor_ID = instructorId;
 
                 Stream str = plan.InputStream;
@@ -197,10 +123,10 @@ namespace SkyExams.Controllers
                     Student tempStudent = db.Students.ToList().Find(s => s.Student_ID == temp.Student_ID);
                     Sys_User tempUser = db.Sys_User.ToList().Find(u => u.SysUser_ID == tempStudent.SysUser_ID);
                     Student_Lesson_Plan newStuPlan = new Student_Lesson_Plan();
-                    newStuPlan.Student_ID = tempUser.SysUser_ID;
+                    newStuPlan.Student_ID = tempStudent.Student_ID;
                     newStuPlan.Lesson_Plan_ID = newPlan.Lesson_Plan_ID;
                     db.Student_Lesson_Plan.Add(newStuPlan);
-                    db.SaveChangesAsync();
+                    db.SaveChanges();
                 }// for each
                 return RedirectToAction("planScreen", new { id = id });
             }// else
@@ -210,6 +136,7 @@ namespace SkyExams.Controllers
         {
             ViewData["loggedId"] = "" + loggedId;
             Lesson_Plan delPlan = db.Lesson_Plan.ToList().Find(p => p.Lesson_Plan_ID == id);
+            ViewData["planeType"] = db.Plane_Type.ToList().Find(p => p.Plane_Type_ID == delPlan.Rating_ID).Type_Description;
             return View(delPlan);
         }// delete Resource
 
@@ -219,8 +146,11 @@ namespace SkyExams.Controllers
             db.Lesson_Plan.Remove(delPlan);
             db.SaveChanges();
             Student_Lesson_Plan delStuPlan = db.Student_Lesson_Plan.ToList().Find(p => p.Lesson_Plan_ID == delPlan.Lesson_Plan_ID);
-            db.Student_Lesson_Plan.Remove(delStuPlan);
-            db.SaveChanges();
+            if(delStuPlan != null)
+            {
+                db.Student_Lesson_Plan.Remove(delStuPlan);
+                db.SaveChanges();
+            }
             return RedirectToAction("planScreen", new { id = loggedId });
         }// delete conformation
